+++ src/main/java/GameState.java
@@ -1,3 +1,4 @@
+import com.doomhamsters.Player;
 import java.util.List;
 
 /**
